pyproject.toml
[project]
name = "calcul-idc-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopandas>=1.0.1",
    "matplotlib>=3.10.0",
    "openpyxl>=3.1.5",
    "plotly>=6.0.1",
    "polars>=1.23.0",
    "pyproj>=3.7.1",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
    "streamlit>=1.42.2",
]
